Proper error handling

+++ project/ui_windows/main_menu.py
@@ -4,24 +4,29 @@ from PySide6.QtCore import Slot
 from PySide6.QtGui import QPixmap, QIcon
 from PySide6.QtWidgets import QMainWindow
 
+from project.classes.exceptions import ContainerNotEmpty, NothingSelected
 from project.classes.userdata import UserData
 from project.classes.recommender import Recommender
 from project.ui.output import Ui_MainMenu
 from project.ui_windows.add_room_window import AddRoomWindow
+from project.ui_windows.confirmation_window import ConfirmationWindow
+from project.ui_windows.error_message_window import ErrorMessageWindow
 from project.ui_windows.room_view_window import RoomViewWindow
 from project.ui_windows.all_plants_window import AllPlantsWindow
 from project.ui_windows.recommendations_window import RecommendationsWindow
 from project.classes.save_and_load_userdata import save_user_data
 from project.ui_windows.notifier_window import NotifierWindow
 from project.classes.notifier import Notifier
+from time import sleep
 
-
-
-import images_qr
+import images_rc
 
 
 class MainMenu(QMainWindow):
-    """Example application"""
+    """
+    The main application window from which all sub windows can be opened.
+    Will save its userdata when closed
+    """
 
     def __init__(self, userdata: UserData) -> None:
         super().__init__()
@@ -59,23 +64,51 @@ class MainMenu(QMainWindow):
         """
         Opens AddRoomWindow that lets the user add a room
         """
-        self.add_room_window = AddRoomWindow(self)
-        self.add_room_window.show()
+        try:
+            self.add_room_window = AddRoomWindow(self)
+        except Exception as e:
+            error_msg = ErrorMessageWindow(e)
+            error_msg.exec()
+        else:
+            self.add_room_window.show()
 
     @Slot()
     def open_room(self) -> None:
         """
         Opens RoomViewWindow that show the spots a room contains.
         """
-        room_name = self.ui.room_list.selectedItems()[0].text()
-        self.room_view_window = RoomViewWindow(room_name , self)
-        self.room_view_window.show()
+        try:
+            selection = self.ui.room_list.selectedItems()
+            if selection:
+                room_name = selection[0].text()
+            else:
+                raise NothingSelected("Selection is empty")
+        except NothingSelected:
+
+            error_msg = ErrorMessageWindow("Please select a room.", "No room selected")
+            error_msg.exec()
+        else:
+            self.room_view_window = RoomViewWindow(room_name , self)
+            self.room_view_window.show()
 
     @Slot()
     def save(self):
-        save_user_data(self.userdata)
+        """
+        Saves user data to json file.
+        Located in %appdata%/Plantfolio/user_data.json when running executable.
+        Located in ./project/userdata.json when run as python project.
+        """
+        try:
+            save_user_data(self.userdata)
+        except Exception as e:
+            error_msg = ErrorMessageWindow(e)
+            error_msg.exec()
 
+    @Slot()
     def water_all(self):
+        """
+        Waters all plants and turns grey for 5 seconds
+        """
         self.userdata.water_all()
         self.update_notifications()
 
@@ -97,17 +130,38 @@ class MainMenu(QMainWindow):
         Deletes the selected room that doesn't contain any spots.
         """
         room_name = room.room_name
-        self.userdata.delete_room(room_name)
-        room.close()
-        self.refresh_rooms()
+        confirm_deletion = ConfirmationWindow(f"Are you sure you want to delete {room_name}?")
+        if confirm_deletion.exec():
+            try:
+                self.userdata.delete_room(room_name)
+            except ContainerNotEmpty:
+                error_msg = ErrorMessageWindow(
+                    f"Room still contains {len(self.userdata.rooms[room_name])} spots. "
+                               "Please remove all spots from the room before deleting",
+                               "Room not empty")
+                error_msg.exec()
+            except Exception as e:
+                error_msg = ErrorMessageWindow(e)
+                error_msg.exec()
+
+
+            else:
+                room.close()
+            finally:
+                self.refresh_rooms()
 
     @Slot()
     def open_notifier(self):
         """
         Opens the notification window
         """
-        self.notifier_window = NotifierWindow(self.notifier)
-        self.notifier_window.show()
+        try:
+            self.notifier_window = NotifierWindow(self.notifier)
+        except Exception as e:
+            error_msg = ErrorMessageWindow(e)
+            error_msg.exec()
+        else:
+            self.notifier_window.show()
 
     def handle_sort_change(self, index):
         """
@@ -139,8 +193,13 @@ class MainMenu(QMainWindow):
         """
         Opens all plants window
         """
-        self.all_plants_window = AllPlantsWindow(self.userdata)
-        self.all_plants_window.show()
+        try:
+            self.all_plants_window = AllPlantsWindow(self.userdata)
+        except Exception as e:
+            error_msg = ErrorMessageWindow(e)
+            error_msg.exec()
+        else:
+            self.all_plants_window.show()
 
     def refresh_rooms(self) -> None:
         """
@@ -155,5 +214,19 @@ class MainMenu(QMainWindow):
         """
         Opens RecommendationsWindow that lets the user see his recommendations
         """
-        self.recommendations_window = RecommendationsWindow(self.recommender, self.userdata)
-        self.recommendations_window.show()
+        try:
+            self.recommendations_window = RecommendationsWindow(self.recommender, self.userdata)
+        except Exception as e:
+            error_msg = ErrorMessageWindow(e)
+            error_msg.exec()
+        else:
+            self.recommendations_window.show()
+
+    def closeEvent(self, event):
+       try:
+           self.save()
+       except Exception as e:
+           error_msg = ErrorMessageWindow(e)
+           error_msg.exec()
+       else:
+           event.accept()
